Zoom Web App Backend/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.1
google-generativeai==0.4.0
firebase-admin==6.5.0
